refactor(entities): extract CopyLinkButton to client component and fix event bubbling
