+++ src/entities/blog/index.ts
@@ -2,4 +2,4 @@ export { PostMainCard } from "./post-main-card";
 export { DateWithReadTime } from "./date-with-read-time";
 export { PostMobileCard } from "./post-mobile-card";
 export { PostDesktopCard } from "./post-desktop-card";
-export { CopyLinkButton } from "./сopy-link-button";
+export { CopyLinkButton } from "./copy-link-button";
